+++ LibTest/js_interop/JSAnyOperatorExtension/equals_t10.dart
@@ -6,7 +6,7 @@
 /// The result of `this == other` in JavaScript.
 ///
 /// @description Check that `equals` returns result of `this == other` in
-/// JavaScript. Test 'null' as `this`.
+/// JavaScript. Test a boolean string  as `this`.
 /// @author sgrekhov22@gmail.com
 
 import 'dart:js_interop';
@@ -14,11 +14,11 @@ import '../../../Utils/expect.dart';
 import 'equals_lib.dart';
 
 void main() {
-  var underTest = null.jsify();
-  testEquals(underTest, "null");
+  var underTest = "true".toJS;
+  testEquals(underTest, "'true'");
   Expect.isTrue(underTest.equals(underTest).dartify());
 
-  underTest = null;
-  testEquals(underTest, "null");
+  underTest = "false".toJS;
+  testEquals(underTest, "'false'");
   Expect.isTrue(underTest.equals(underTest).dartify());
 }
