+++ LibTest/js_interop/JSAnyOperatorExtension/equals_t15.dart
@@ -6,21 +6,19 @@
 /// The result of `this == other` in JavaScript.
 ///
 /// @description Check that `equals` returns result of `this == other` in
-/// JavaScript. Test a function as `this`.
+/// JavaScript. Test a boolean as `this`.
 /// @author sgrekhov22@gmail.com
 
 import 'dart:js_interop';
-import 'dart:js_interop_unsafe';
 import '../../../Utils/expect.dart';
 import 'equals_lib.dart';
-import '../js_utils.dart';
 
 void main() {
-  eval(r'''
-    function bar() {}
-  ''');
+  var underTest = true.toJS;
+  testEquals(underTest, "true");
+  Expect.isTrue(underTest.equals(underTest).dartify());
 
-  var underTest = globalContext["bar"].jsify();
-  testEquals(underTest, "bar");
+  underTest = false.toJS;
+  testEquals(underTest, "false");
   Expect.isTrue(underTest.equals(underTest).dartify());
 }
